+++ src/Options/HandlerOptions.php
@@ -13,6 +13,11 @@ final class HandlerOptions extends AbstractOptions
     {
         return $this->get('formatter', false);
     }
+
+    public function getProcessors() : array
+    {
+        return $this->get('processors', []);
+    }
     
     protected static function validate(array $options)
     {
@@ -21,6 +26,20 @@ final class HandlerOptions extends AbstractOptions
             
             if (! ($formatter instanceof FormatterInterface || is_array($formatter))) {
                 throw InvalidOptionsException::forInvalidFormatter($formatter);
+            }
+        }
+
+        if (array_key_exists('processors', $options)) {
+            $processors = $options['processors'];
+
+            if (! is_array($processors)) {
+                throw InvalidOptionsException::forInvalidProcessors($options['processors']);
+            }
+
+            foreach ($processors as $processor) {
+                if (! (is_callable($processor) || is_array($processor))) {
+                    throw InvalidOptionsException::forInvalidProcessor($processor);
+                }
             }
         }
     }